+++ lib/Core/SubscriptionChange/SubscriptionChangesReader.php
@@ -13,6 +13,7 @@ class SubscriptionChangesReader {
 	 */
 	public static function mapToSubscriptionsChanges(array $urls, bool $subscribed): array {
 		$subscriptionChanges = [];
+        $urls = array_filter($urls, function(string $url) {return filter_var($url, FILTER_VALIDATE_URL) !== false; });
 		foreach ($urls as $url) {
 			$subscriptionChanges[] = new SubscriptionChange($url, $subscribed);
 		}