+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## upcoming
+### Changed
+- Ignore subscriptions that have no url
+
 ## 3.1.0 - 2021-10-18
 ### Changed
 - Add timestamp to subscription change response @JonOfUs